xhci: Remove xhci_handshake_check_state API
Because of commit f4a5311dfd1cbf9440f0 ("xhci: make xhci_handshake timeout
for xhci_reset() adjustable"), the reset timeout is now back to a sane
default and also became configurable: in turn it should avoid making the
watchdog timeout again.

All of this made the CAF-issued xhci_handshake_check_state() redundant,
therefore it is best to remove it.

Fixes: https://github.com/miyoubi/sodp_kernel_sony_sdm845/commit/089fc989d534f082c865129d8a1eea8cfa29894b ("usb: XHCI: Implement xhci_handshake_check_state() API")
Signed-off-by: voidanix <voidanix@keyedlimepie.org>
Signed-off-by: chickendrop89 <chickendrop89@gmail.com>